feat: add optional fields for info page
