+++ src/modules/items/index.ts
@@ -1,5 +1,5 @@
 export { Items } from './components/Items'
-export { ItemInfo } from './components/ItemInfo'
+export { ItemInfo } from './components/itemInfo/ItemInfo'
 export { SearchItems } from './components/SearchItems'
 export { LatestItems } from './components/LatestItems'
 export { itemsReducer } from './store/slice'
