+++ src/shared/localization/en.json
@@ -45,7 +45,15 @@
   "app.registration-page.link": "Login",
   "app.create.header": "Create collection",
   "app.create.title": "Title",
+  "app.create.description": "Description",
+  "app.create.tag": "Tags",
+  "app.create.tag-added": "Tag already added!",
   "app.create.theme": "Theme",
+  "app.create.theme-Alcohol": "Alcohol",
+  "app.create.theme-Books": "Books",
+  "app.create.theme-Movies": "Movies",
+  "app.create.theme-Music": "Music",
+  "app.create.theme-Games": "Games",
   "app.create.button": "Create",
   "app.edit.button": "Save",
   "app.edit.header": "Edit item",
@@ -71,5 +79,8 @@
   "app.checkbox.status": "Status",
   "app.checkbox.terminated": "Terminated",
   "app.checkbox.original": "Original",
+  "app.checkbox.not-original": "Not original",
+  "app.checkbox.use-status": "Available",
+  "app.checkbox.not-status": "Not available",
   "app.components.back-button": "Back"
 }
