+++ src/shared/localization/ru.json
@@ -45,7 +45,15 @@
   "app.registration-page.link": "Войти",
   "app.create.header": "Создать коллекцию",
   "app.create.title": "Название",
+  "app.create.description": "Описание",
+  "app.create.tag": "Теги",
+  "app.create.tag-added": "Тег уже добавлен!",
   "app.create.theme": "Тема",
+  "app.create.theme-Alcohol": "Алкоголь",
+  "app.create.theme-Books": "Книги",
+  "app.create.theme-Movies": "Фильмы",
+  "app.create.theme-Music": "Музыка",
+  "app.create.theme-Games": "Игры",
   "app.create.button": "Создать",
   "app.edit.button": "Сохранить",
   "app.edit.header": "Изменить коллекцию",
@@ -71,5 +79,8 @@
   "app.checkbox.status": "Статус",
   "app.checkbox.terminated": "Прекращено",
   "app.checkbox.original": "Оригинал",
+  "app.checkbox.not-original": "Не оригинал",
+  "app.checkbox.use-status": "Доступен",
+  "app.checkbox.not-status": "Недоступен",
   "app.components.back-button": "Назад"
 }
